Split object inspector test file
